% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{DataInfoIE}
\alias{DataInfoIE}
\title{Data information for reference samples}
\usage{
DataInfoIE(data, rho)
}
\arguments{
\item{data}{data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors.}

\item{rho}{the sampling fraction can be input as a vector for each assemblage or a numeric common value for all assemblages.}
}
\value{
a data.frame including assemblage name (\code{Assemblage}), sample size (\code{n}), 
total abundance (\code{N}), sampling fraction (\code{rho}), observed species richness (\code{S.obs}), 
sample coverage estimates of the reference sample (\code{SC(n)}), sample coverage estimate for twice the reference sample size (\code{SC(2n)}),
the first five species abundance counts (\code{f1}--\code{f5}).\cr
}
\description{
\code{DataInfoIE} provides basic data information for diversity based on a reference sample.
}
\examples{
DataInfoIE(spider, rho = 0.3)


}
