% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{estimateIE}
\alias{estimateIE}
\title{Compute diversity estimates with a particular set of sample sizes/coverages}
\usage{
estimateIE(
  data,
  rho,
  q = c(0, 1, 2),
  base = "coverage",
  level = NULL,
  nboot = 50,
  conf = 0.95
)
}
\arguments{
\item{data}{data can be input as a vector of species abundances (for a single assemblage), matrix/data.frame (species by assemblages), or a list of species abundance vectors.}

\item{rho}{the sampling fraction can be input as a vector for each assemblage or a numeric common value for all assemblages.}

\item{q}{a numerical vector specifying the diversity orders. Default is \code{c(0, 1, 2)}.}

\item{base}{selection of sample-size-based (\code{base = "size"}) or coverage-based (\code{base = "coverage"}) rarefaction and extrapolation.}

\item{level}{A numerical vector specifying the particular sample sizes or sample coverages (between 0 and 1) for which diversity estimates (q = 0, 1 and 2) will be computed. \cr
If \code{base = "coverage"} (default) and \code{level = NULL}, then this function computes the diversity estimates for the minimum sample coverage among all samples extrapolated to double reference sizes or total individuals. \cr
If \code{base = "size"} and \code{level = NULL}, then this function computes the diversity estimates for the minimum sample size among all samples extrapolated to double reference sizes or total individuals.}

\item{nboot}{a positive integer specifying the number of bootstrap replications when assessing sampling uncertainty and constructing confidence intervals. Enter 0 to skip the bootstrap procedures. Default is 50.}

\item{conf}{a positive number < 1 specifying the level of confidence interval. Default is 0.95.}
}
\value{
a data.frame of diversity table including the following arguments: (when \code{base = "coverage"})
\item{Assemblage}{the name of assemblage.}
\item{Order.q}{the diversity order of q.}
\item{SC}{the target standardized coverage value.}
\item{m}{the corresponding sample size for the standardized coverage value.}
\item{Method}{Rarefaction, Observed, or Extrapolation, depending on whether the target coverage is less than, equal to, or greater than the coverage of the reference sample.}
\item{qIE}{the estimated diversity of order q for the target coverage value. The estimate for complete coverage (when \code{base = "coverage"} and \code{level = 1}, or \code{rho = 1}) represents the estimated asymptotic diversity.}
\item{s.e.}{standard error of diversity estimate.}
\item{qIE.LCL and qIE.UCL}{the bootstrap lower and upper confidence limits for the diversity of order q at the specified level (with a default value of 0.95).}
Similar output is obtained for \code{base = "size"}. \cr\cr
}
\description{
\code{estimateIE} computes diversity with a particular set of user-specified levels of sample sizes or sample coverages. If no sample sizes or coverages are specified, this function by default computes diversity estimates for the minimum sample coverage or minimum sample size among all samples extrapolated to double reference sizes or total individuals.
}
\examples{
data("spider")
output_est_cov <- estimateIE(spider, rho = 0.3, q = c(0, 1, 2), 
                             base = "coverage", level = c(0.94, 0.96))
output_est_cov

output_est_size <- estimateIE(spider, rho = 0.3, q = c(0, 1, 2),
                              base = "size", level = c(150, 250))
output_est_size


}
