% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{ggiNEXTIE}
\alias{ggiNEXTIE}
\title{ggplot2 extension for an iNEXTIE object}
\usage{
ggiNEXTIE(output, type = 1:3)
}
\arguments{
\item{output}{an \code{iNEXTIE} object computed by \code{iNEXTIE}.}

\item{type}{three types of plots: sample-size-based rarefaction/extrapolation curve (\code{type = 1}); 
sample completeness curve (\code{type = 2}); coverage-based rarefaction/extrapolation curve (\code{type = 3}).}
}
\value{
a \code{ggplot2} object for sample-size-based rarefaction/extrapolation curve (\code{type = 1}), sample completeness curve (\code{type = 2}), and coverage-based rarefaction/extrapolation curve (\code{type = 3}).
}
\description{
\code{ggiNEXTIE} is a \code{ggplot} extension for an \code{iNEXTIE} object to plot sample-size- and coverage-based rarefaction/extrapolation sampling curves along with a bridging sample completeness curve.
}
\examples{
# Plot three types of curves of diversity for abundance data with order q = 0, 1, 2
set.seed(2025)
data = list("Site_1" = as.numeric(rmultinom(1, 200, 1:30)), 
            "Site_2" = as.numeric(rmultinom(1, 300, 1:40)))
output_iNEXT <- iNEXTIE(data, rho = 0.3, q = c(0, 1, 2))
ggiNEXTIE(output_iNEXT)


}
