% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainFun.r
\name{print.iNEXTIE}
\alias{print.iNEXTIE}
\title{Printing iNEXTIE object}
\usage{
\method{print}{iNEXTIE}(x, ...)
}
\arguments{
\item{x}{an \code{iNEXTIE} object computed by \code{iNEXTIE}.}

\item{...}{additional arguments.}
}
\value{
a list of three objects (see \code{iNEXTIE} for more details) with simplified outputs and notes.
}
\description{
\code{print.iNEXTIE}: Print method for objects inheriting from class "iNEXTIE"
}
