\name{spider}
\alias{spider}
\docType{data}
\title{
  Spider abundance data
}
\description{
  Spider abundance data (Sackett et al., 2011) were collected from experimental plots in the Harvard Forest, located in Petersham, Massachusetts, USA. The experiment involved two canopy-level manipulations designed to mimic structural changes caused either by logging or infestation by the hemlock woolly adelgid. The study aimed to assess the long-term ecological consequences of losing the dominant forest tree, eastern hemlock (Tsuga canadensis), due to infestation by the nonnative insect Adelges tsugae (Ellison et al., 2010). Two treatments were applied: (1) the hemlock-girdled treatment, where bark and cambium were removed using knives or chainsaws, and trees were left standing to die gradually over 30 months, mimicking mortality due to adelgid infestation; and
(2) the hemlock-logged treatment, where all hemlocks greater than 20 cm in diameter were harvested and removed, along with any merchantable white pine and hardwoods.
}
\usage{data("spider")}
\format{
  \code{spider} is a list of two assemblages vector data. \cr
  $ Girdled: num [1:26] 46 22 17 15 15 9 8 6 6 4 ... \cr
  $ Logged : num [1:37] 88 22 16 15 13 10 8 8 7 7 ... \cr
}
\source{
Ellison, A. M., Barker‐Plotkin, A. A., Foster, D. R., and Orwig, D. A. (2010). Experimentally testing the role of foundation species in forests: the Harvard Forest Hemlock Removal Experiment. Methods in Ecology and Evolution, 1(2), 168-179. \cr\cr
Sackett, T. E., Record, S., Bewick, S., Baiser, B., Sanders, N. J., and Ellison, A. M. (2011). Response of macroarthropod assemblages to the loss of hemlock (Tsuga canadensis), a foundation species. Ecosphere, 2(7), 1-16. \cr\cr
}
\examples{
  data(spider)
}
\keyword{datasets}
